fastapi-app/requirements.txt
fastapi
uvicorn[standard]
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-logging
opentelemetry-distro